requirements.txt
psycopg2-binary>=2.9.0
duckdb>=0.9.0
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0
scikit-learn>=1.1.0
lightgbm>=4.0.0
shap>=0.41.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
